refactor: moved schema and other form improvements
